Changed to use mcuboot version 1.8.0

+++ boot/nuttx/src/flash_map_backend/flash_map_backend.c
@@ -421,7 +421,7 @@ int flash_area_read(const struct flash_area *fa, uint32_t off,
   /* Reposition the file offset from the beginning of the flash area */
 
   seekpos = lseek(dev->fd, (off_t)off, SEEK_SET);
-  if (seekpos != (off_t)off)
+  if (seekpos != off)
     {
       int errcode = errno;
 
@@ -489,7 +489,7 @@ int flash_area_write(const struct flash_area *fa, uint32_t off,
   /* Reposition the file offset from the beginning of the flash area */
 
   seekpos = lseek(dev->fd, (off_t)off, SEEK_SET);
-  if (seekpos != (off_t)off)
+  if (seekpos != off)
     {
       int errcode = errno;
 
@@ -595,15 +595,15 @@ int flash_area_erase(const struct flash_area *fa, uint32_t off, uint32_t len)
  *
  ****************************************************************************/
 
-uint32_t flash_area_align(const struct flash_area *fa)
+uint8_t flash_area_align(const struct flash_area *fa)
 {
   /* MTD access alignment is handled by the character and block device
    * drivers.
    */
 
-  const uint32_t minimum_write_length = 1;
+  const uint8_t minimum_write_length = 1;
 
-  BOOT_LOG_INF("ID:%" PRIu8 " align:%" PRIu32,
+  BOOT_LOG_INF("ID:%" PRIu8 " align:%" PRIu8,
                fa->fa_id, minimum_write_length);
 
   return minimum_write_length;
